+++ src/gui/dialogs.rs
@@ -4,8 +4,10 @@ mod lua_color;
 mod lua_execute;
 mod lua_fill;
 mod pattern_fill;
+mod truncate;
 
 pub use {
     auto_save_reload::AutoSaveReloadDialog, jump::JumpDialog, lua_color::LuaColorDialog,
     lua_execute::LuaExecuteDialog, lua_fill::LuaFillDialog, pattern_fill::PatternFillDialog,
+    truncate::TruncateDialog,
 };