+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - About dialog with version info + links
 - Clickable file size label in bottom right corner
+- Functionality to change the length of the data (truncate/extend)
 
 ### Changed